Separate parsers to parsers.py
